Workaround for Cabal ticket #944.
This patch is a modified version of one submitted 
by Paul van der Walt.
It adds a "haskeline_" prefix to all of Haskeline's
C symbol names.  This avoids clashing symbol names with other
packages that implement the same C functions.  
(In particular, it prevents a clash with vty on "mk_wcswidth".)