Refactor terminal unit tests.
Previously we used the "script" command-line utility to capture output from
programs that expected to be run in a terminal.  
Now, we use openpty and other C function (as wrapped by the unix library) to
run them directly from Haskell.  

As a related change, we now return the individual chunks of bytes that were
output from the process that we're testing.  This lets us test its behavior
more specifically, in particular in the "file style" tests with an incomplete
byte sequence at EOF.