Convert to simple build-type
The logic in Setup.hs doesn't affect the build per-se, but rather
prints out a purely informational warning on non-windows systems
in case `terminfo` is disabled. However, the upcoming Cabal 2.0 API
breaks the Setup.hs script so it's easier to just convert to simple
build-type and avoid the additional complexity and overhead.