Preferring fmap or <$> to liftM and useKey to +>, defining flipDir and doAfter
Regarding preferring `useKey` over `+>`, the rationale is: while `>+>`
is truly just a generic combinator (it is indeed just `flip (fmap . flip
(>=>))`), so giving it a name is complicated, `useKey` is not, and the
operator `+>` is just one more thing to remember. `useKey` is way more
informative and can be used infixed anyway; the only drawback is that
some parenthesis are needed.

Renamed killAndStoreIE to killAndStoreE as I didn't know what "I" stood
for, and killAndStoreCmd to killAndStoreC for consistency.

Moved copyAndStore out for storing.